Implemented Day 4 successfully

+++ day-1/src/main.rs
@@ -1,7 +1,7 @@
 use inputs::get_input;
 
-fn part1(inputs: &Vec<i32>) -> Result<i32, String> {
-    let calculated: Vec<i32> = inputs.into_iter().map(|i| 2020 - i).collect();
+fn part1(inputs: &Vec<u64>) -> Result<u64, String> {
+    let calculated: Vec<u64> = inputs.into_iter().map(|i| 2020 - i).collect();
 
     for num in inputs {
         if calculated.iter().any(|i| i == num) {
@@ -12,8 +12,8 @@ fn part1(inputs: &Vec<i32>) -> Result<i32, String> {
     Err(String::from("Could not find match in inputs!"))
 }
 
-fn part2(inputs: &Vec<i32>) -> Result<i32, String> {
-    let calculated: Vec<i32> = inputs.into_iter().map(|i| 2020 - i).collect();
+fn part2(inputs: &Vec<u64>) -> Result<u64, String> {
+    let calculated: Vec<u64> = inputs.into_iter().map(|i| 2020 - i).collect();
 
     for num in inputs {
         for num2 in inputs {
@@ -29,7 +29,7 @@ fn part2(inputs: &Vec<i32>) -> Result<i32, String> {
 }
 
 fn main() {
-    let numbers = get_input::<i32>("inputs/day-1.txt").expect("Could not parse path!");
+    let numbers = get_input::<u64>("inputs/day-1.txt").expect("Could not parse path!");
 
     println!("Part 1 solution: {:?}", part1(&numbers));
     println!("Part 1 solution: {:?}", part2(&numbers));
